+++ test/inject-loggedUser-koa1-api-key-microservice-token.spec.ts
@@ -1,4 +1,5 @@
-import Koa2 from "koa";
+// @ts-ignore
+import Koa1 from "koa1";
 import nock from 'nock';
 import chai, { expect } from 'chai';
 import { RWAPIMicroservice } from 'main';
@@ -8,11 +9,12 @@ import bunyan from "bunyan";
 import type Koa from "koa";
 import Router from "koa-router";
 import koaBody from "koa-body";
+import convert from "koa-convert";
 import type { Server } from "http";
 import type Request from "superagent";
 import constants from './utils/test.constants';
 import ChaiHttp from 'chai-http';
-import { mockValidateRequestWithApiKeyAndUserToken } from "./utils/mocks";
+import { mockValidateRequestWithApiKeyAndUserToken, mockValidateRequestWithUserToken } from "./utils/mocks";
 import { mockCloudWatchLogRequestsSequence } from "../src/test-mocks";
 import { BootstrapArguments } from "../src/types";
 
@@ -24,12 +26,12 @@ nock.enableNetConnect('127.0.0.1');
 
 let requester: ChaiHttp.Agent;
 
-describe('Injecting logged user data - Koa v2.x with API key and user token', () => {
+describe('Injecting logged user data - Koa v1.x with API key and MICROSERVICE token', () => {
 
     before(nock.cleanAll);
 
-    it('GET request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
-        const app: Koa = new Koa2();
+    it('GET request with a JWT microserviceToken in the header should fetch the microservice data and pass it along as loggedUser (happy case)', async () => {
+        const app: Koa = new Koa1();
 
         const logger: Logger = bunyan.createLogger({
             name: 'logger name',
@@ -40,10 +42,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -57,17 +63,17 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.get('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
-            expect(ctx.request.query).to.have.property('loggedUser').and.equal(JSON.stringify(constants.USER));
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
+            expect(ctx.request.query).to.have.property('loggedUser').and.equal(JSON.stringify(constants.MICROSERVICE));
             ctx.body = 'ok';
         });
 
-        app.use(koaBody());
-        app.use(RWAPIMicroservice.bootstrap(registerOptions));
+        app.use(convert.back(koaBody()));
+        app.use(convert.back(RWAPIMicroservice.bootstrap(registerOptions)));
 
         app
-            .use(testRouter.routes())
-            .use(testRouter.allowedMethods());
+            .use(convert.back(testRouter.routes()))
+            .use(convert.back(testRouter.allowedMethods()));
 
         const server: Server = app.listen(3010);
 
@@ -75,15 +81,15 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .get('/test')
-            .set('x-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`);
 
         response.status.should.equal(200);
         response.text.should.equal('ok');
     });
 
-    it('DELETE request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
-        const app: Koa = new Koa2();
+    it('DELETE request with a JWT microserviceToken in the header should fetch the microservice data and pass it along as loggedUser (happy case)', async () => {
+        const app: Koa = new Koa1();
 
         const logger: Logger = bunyan.createLogger({
             name: 'logger name',
@@ -94,10 +100,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -111,17 +121,17 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.delete('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
-            expect(ctx.request.query).to.have.property('loggedUser').and.equal(JSON.stringify(constants.USER));
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
+            expect(ctx.request.query).to.have.property('loggedUser').and.equal(JSON.stringify(constants.MICROSERVICE));
             ctx.body = 'ok';
         });
 
-        app.use(koaBody());
-        app.use(RWAPIMicroservice.bootstrap(registerOptions));
+        app.use(convert.back(koaBody()));
+        app.use(convert.back(RWAPIMicroservice.bootstrap(registerOptions)));
 
         app
-            .use(testRouter.routes())
-            .use(testRouter.allowedMethods());
+            .use(convert.back(testRouter.routes()))
+            .use(convert.back(testRouter.allowedMethods()));
 
         const server: Server = app.listen(3010);
 
@@ -129,15 +139,16 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .delete('/test')
-            .set('X-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            .set('x-api-key', `api-key-test`)
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
+
 
         response.status.should.equal(200);
         response.text.should.equal('ok');
     });
 
-    it('POST request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
-        const app: Koa = new Koa2();
+    it('POST request with a JWT microserviceToken in the header should fetch the microservice data and pass it along as loggedUser (happy case)', async () => {
+        const app: Koa = new Koa1();
 
         const logger: Logger = bunyan.createLogger({
             name: 'logger name',
@@ -148,10 +159,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -165,18 +180,18 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.post('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
             ctx.request.body.should.have.property('data').and.deep.equal('test');
-            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.USER);
+            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.MICROSERVICE);
             ctx.body = 'ok';
         });
 
-        app.use(koaBody());
-        app.use(RWAPIMicroservice.bootstrap(registerOptions));
+        app.use(convert.back(koaBody()));
+        app.use(convert.back(RWAPIMicroservice.bootstrap(registerOptions)));
 
         app
-            .use(testRouter.routes())
-            .use(testRouter.allowedMethods());
+            .use(convert.back(testRouter.routes()))
+            .use(convert.back(testRouter.allowedMethods()));
 
         const server: Server = app.listen(3010);
 
@@ -184,8 +199,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .post('/test')
-            .set('X-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`)
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`)
             .send({
                 data: 'test'
             });
@@ -194,8 +209,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
         response.text.should.equal('ok');
     });
 
-    it('PATCH request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
-        const app: Koa = new Koa2();
+    it('PATCH request with a JWT microserviceToken in the header should fetch the microservice data and pass it along as loggedUser (happy case)', async () => {
+        const app: Koa = new Koa1();
 
         const logger: Logger = bunyan.createLogger({
             name: 'logger name',
@@ -206,10 +221,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -223,18 +242,18 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.patch('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
             ctx.request.body.should.have.property('data').and.deep.equal('test');
-            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.USER);
+            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.MICROSERVICE);
             ctx.body = 'ok';
         });
 
-        app.use(koaBody());
-        app.use(RWAPIMicroservice.bootstrap(registerOptions));
+        app.use(convert.back(koaBody()));
+        app.use(convert.back(RWAPIMicroservice.bootstrap(registerOptions)));
 
         app
-            .use(testRouter.routes())
-            .use(testRouter.allowedMethods());
+            .use(convert.back(testRouter.routes()))
+            .use(convert.back(testRouter.allowedMethods()));
 
         const server: Server = app.listen(3010);
 
@@ -242,8 +261,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .patch('/test')
-            .set('X-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`)
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`)
             .send({
                 data: 'test'
             });
@@ -252,8 +271,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
         response.text.should.equal('ok');
     });
 
-    it('PUT request with a JWT microserviceToken in the header should fetch the user data and pass it along as loggedUser (happy case)', async () => {
-        const app: Koa = new Koa2();
+    it('PUT request with a JWT microserviceToken in the header should fetch the microservice data and pass it along as loggedUser (happy case)', async () => {
+        const app: Koa = new Koa1();
 
         const logger: Logger = bunyan.createLogger({
             name: 'logger name',
@@ -264,10 +283,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -281,18 +304,18 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.put('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
             ctx.request.body.should.have.property('data').and.deep.equal('test');
-            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.USER);
+            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.MICROSERVICE);
             ctx.body = 'ok';
         });
 
-        app.use(koaBody());
-        app.use(RWAPIMicroservice.bootstrap(registerOptions));
+        app.use(convert.back(koaBody()));
+        app.use(convert.back(RWAPIMicroservice.bootstrap(registerOptions)));
 
         app
-            .use(testRouter.routes())
-            .use(testRouter.allowedMethods());
+            .use(convert.back(testRouter.routes()))
+            .use(convert.back(testRouter.allowedMethods()));
 
         const server: Server = app.listen(3010);
 
@@ -300,8 +323,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .put('/test')
-            .set('X-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`)
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`)
             .send({
                 data: 'test'
             });