+++ test/inject-loggedUser-koa1-api-key-user-token.spec.ts
@@ -14,9 +14,7 @@ import type { Server } from "http";
 import type Request from "superagent";
 import constants from './utils/test.constants';
 import ChaiHttp from 'chai-http';
-import {
-    mockValidateRequestWithApiKeyAndUserToken
-} from "./utils/mocks";
+import { mockValidateRequestWithApiKeyAndUserToken } from "./utils/mocks";
 import { mockCloudWatchLogRequestsSequence } from "../src/test-mocks";
 import { BootstrapArguments } from "../src/types";
 
@@ -44,7 +42,7 @@ describe('Injecting logged user data - Koa v1.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken('https://controltower.dev');
+        mockValidateRequestWithApiKeyAndUserToken({ gatewayUrl: 'https://controltower.dev' });
         mockCloudWatchLogRequestsSequence({
             application: constants.APPLICATION,
             user: constants.USER,
@@ -99,7 +97,7 @@ describe('Injecting logged user data - Koa v1.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken('https://controltower.dev');
+        mockValidateRequestWithApiKeyAndUserToken({ gatewayUrl: 'https://controltower.dev' });
         mockCloudWatchLogRequestsSequence({
             application: constants.APPLICATION,
             user: constants.USER,
@@ -155,7 +153,7 @@ describe('Injecting logged user data - Koa v1.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken('https://controltower.dev');
+        mockValidateRequestWithApiKeyAndUserToken({ gatewayUrl: 'https://controltower.dev' });
         mockCloudWatchLogRequestsSequence({
             application: constants.APPLICATION,
             user: constants.USER,
@@ -214,7 +212,7 @@ describe('Injecting logged user data - Koa v1.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken('https://controltower.dev');
+        mockValidateRequestWithApiKeyAndUserToken({ gatewayUrl: 'https://controltower.dev' });
         mockCloudWatchLogRequestsSequence({
             application: constants.APPLICATION,
             user: constants.USER,
@@ -273,7 +271,7 @@ describe('Injecting logged user data - Koa v1.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken('https://controltower.dev');
+        mockValidateRequestWithApiKeyAndUserToken({ gatewayUrl: 'https://controltower.dev' });
         mockCloudWatchLogRequestsSequence({
             application: constants.APPLICATION,
             user: constants.USER,