+++ test/inject-loggedUser-koa1-user-token.spec.ts
@@ -2,7 +2,7 @@
 import Koa1 from "koa1";
 import nock from 'nock';
 import chai, { expect } from 'chai';
-import { BootstrapArguments, RWAPIMicroservice } from 'main';
+import { RWAPIMicroservice } from 'main';
 import type Logger from "bunyan";
 import type { LogLevel } from "bunyan";
 import bunyan from "bunyan";
@@ -18,6 +18,7 @@ import {
     mockValidateRequestWithUserToken
 } from "./utils/mocks";
 import { mockCloudWatchLogRequestsSequence } from "../src/test-mocks";
+import { BootstrapArguments } from "../src/types";
 
 chai.should();
 chai.use(ChaiHttp);