+++ test/inject-loggedUser-koa2-api-key-microservice-token.spec.ts
@@ -1,3 +1,4 @@
+import type Koa from "koa";
 import Koa2 from "koa";
 import nock from 'nock';
 import chai, { expect } from 'chai';
@@ -5,14 +6,13 @@ import { RWAPIMicroservice } from 'main';
 import type Logger from "bunyan";
 import type { LogLevel } from "bunyan";
 import bunyan from "bunyan";
-import type Koa from "koa";
 import Router from "koa-router";
 import koaBody from "koa-body";
 import type { Server } from "http";
 import type Request from "superagent";
 import constants from './utils/test.constants';
 import ChaiHttp from 'chai-http';
-import { mockValidateRequestWithApiKeyAndUserToken } from "./utils/mocks";
+import { mockValidateRequestWithApiKeyAndUserToken, mockValidateRequestWithUserToken } from "./utils/mocks";
 import { mockCloudWatchLogRequestsSequence } from "../src/test-mocks";
 import { BootstrapArguments } from "../src/types";
 
@@ -24,7 +24,7 @@ nock.enableNetConnect('127.0.0.1');
 
 let requester: ChaiHttp.Agent;
 
-describe('Injecting logged user data - Koa v2.x with API key and user token', () => {
+describe('Injecting logged user data - Koa v2.x with API key and MICROSERVICE token', () => {
 
     before(nock.cleanAll);
 
@@ -40,10 +40,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -57,8 +61,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.get('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
-            expect(ctx.request.query).to.have.property('loggedUser').and.equal(JSON.stringify(constants.USER));
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
+            expect(ctx.request.query).to.have.property('loggedUser').and.equal(JSON.stringify(constants.MICROSERVICE));
             ctx.body = 'ok';
         });
 
@@ -75,8 +79,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .get('/test')
-            .set('x-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`);
 
         response.status.should.equal(200);
         response.text.should.equal('ok');
@@ -94,10 +98,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -111,8 +119,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.delete('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
-            expect(ctx.request.query).to.have.property('loggedUser').and.equal(JSON.stringify(constants.USER));
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
+            expect(ctx.request.query).to.have.property('loggedUser').and.equal(JSON.stringify(constants.MICROSERVICE));
             ctx.body = 'ok';
         });
 
@@ -129,8 +137,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .delete('/test')
-            .set('X-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`);
 
         response.status.should.equal(200);
         response.text.should.equal('ok');
@@ -148,10 +156,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -165,9 +177,9 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.post('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
             ctx.request.body.should.have.property('data').and.deep.equal('test');
-            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.USER);
+            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.MICROSERVICE);
             ctx.body = 'ok';
         });
 
@@ -184,8 +196,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .post('/test')
-            .set('X-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`)
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`)
             .send({
                 data: 'test'
             });
@@ -206,10 +218,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -223,9 +239,9 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.patch('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
             ctx.request.body.should.have.property('data').and.deep.equal('test');
-            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.USER);
+            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.MICROSERVICE);
             ctx.body = 'ok';
         });
 
@@ -242,8 +258,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .patch('/test')
-            .set('X-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`)
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`)
             .send({
                 data: 'test'
             });
@@ -264,10 +280,14 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
             }],
         });
 
-        mockValidateRequestWithApiKeyAndUserToken({gatewayUrl: 'https://controltower.dev'});
+        mockValidateRequestWithApiKeyAndUserToken({
+            gatewayUrl: 'https://controltower.dev',
+            user: constants.MICROSERVICE,
+            token: constants.MICROSERVICE_TOKEN
+        });
         mockCloudWatchLogRequestsSequence({
-            user: constants.USER,
-            application: constants.APPLICATION,
+            user: constants.MICROSERVICE,
+            application: constants.APPLICATION
         });
 
         const registerOptions: BootstrapArguments = {
@@ -281,9 +301,9 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const testRouter: Router = new Router();
         testRouter.put('/test', (ctx: Koa.Context) => {
-            ctx.request.should.have.header('Authorization', `Bearer ${constants.USER_TOKEN}`);
+            ctx.request.should.have.header('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`);
             ctx.request.body.should.have.property('data').and.deep.equal('test');
-            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.USER);
+            ctx.request.body.should.have.property('loggedUser').and.deep.equal(constants.MICROSERVICE);
             ctx.body = 'ok';
         });
 
@@ -300,8 +320,8 @@ describe('Injecting logged user data - Koa v2.x with API key and user token', ()
 
         const response: Request.Response = await requester
             .put('/test')
-            .set('X-api-key', `api-key-test`)
-            .set('Authorization', `Bearer ${constants.USER_TOKEN}`)
+            .set('Authorization', `Bearer ${constants.MICROSERVICE_TOKEN}`)
+            .set('x-api-key', `api-key-test`)
             .send({
                 data: 'test'
             });