+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v5.0.0-rc3
+
+- Modify `requireAPIKey` logic.
+- Add API Key validation logic when making calls to `requestToMicroservice`.
+
 # v5.0.0-rc1
 
 - Make application logging to AWS CloudWatch more flexible.